ai/requirements.txt
python-dotenv>=1.0,<2
flask>=3.0,<4
gunicorn>=23,<24
httpx>=0.27,<0.29
pydantic>=2.7,<3
jinja2>=3.1,<4
yfinance>=0.2.40,<0.3
google-genai>=1.0,<2
langgraph>=0.2.45,<0.7
